split out DetObjectFunc as the solution for non-dettyped detectors for LCLS1 detectors now broke LCLS2 psana
