+++ smalldata_tools/ana_funcs/droplet2Photons.py
@@ -7,7 +7,7 @@ import skimage.measure as measure
 import scipy.ndimage.filters as filters
 from scipy import sparse
 import time
-from smalldata_tools.DetObject import DetObjectFunc
+from smalldata_tools.DetObjectFunc import DetObjectFunc
 
 class droplet2Photons(DetObjectFunc):
     '''
